Backend/requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2024.2
sqlparse==0.5.3
